# Faster dm fidelity which is compatible with GPU