We use https://arxiv.org/abs/2211.02623 to rewrite the fidelity as

$$
F(\rho,\sigma)=\mathrm{Tr}\left[\sqrt{\sqrt\rho \sigma \sqrt\rho}\right]^2=\mathrm{Tr}[\sqrt{\rho\sigma}]^2.
$$

This is more efficient numerically, and it also fixes the bug on GPUs (that don't support `jax.scipy.linalg.sqrtm`).

Closes https://linear.app/dynamiqs/issue/DYN-105/fix-dm-fidelity.

Speed test on morbier:
```python
import jax
import dynamiqs as dq

for n in [8, 128, 1024]:
    print(f'=== n={n}')
    key = jax.random.PRNGKey(42)
    k1, k2 = jax.random.split(key)
    x = dq.rand.dm(k1, (n, n))
    y = dq.rand.dm(k2, (n, n))

    print('no jit > ', end='')
    %timeit dq.fidelity(x, y).block_until_ready()
    jfid = jax.jit(dq.fidelity)
    jfid(x, y).block_until_ready()
    print('   jit > ', end='')
    %timeit jfid(x, y).block_until_ready()
```
Previous version:
```
=== n=8
no jit > 285 µs ± 834 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
   jit > 18.5 µs ± 29.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
=== n=128
no jit > 12.8 ms ± 109 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
   jit > 12.6 ms ± 90.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
=== n=1024
no jit > 1.16 s ± 3.17 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
   jit > 1.08 s ± 8.64 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
````
New version:
```
=== n=8
no jit > 193 µs ± 118 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
   jit > 12.1 µs ± 11.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
=== n=128
no jit > 4.74 ms ± 23.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
   jit > 4.46 ms ± 23.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
=== n=1024
no jit > 472 ms ± 4.59 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
   jit > 475 ms ± 2.32 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```